Fix "stuck" long cycle modes
The trigger at time for widget alarms was being set relative to the moment setup was called, and that's fine as long as the alarm goes off before the device is turned off. If a user sets a repeating alarm for longer than a day, but turns off their device every day, the alarm would never go off since it would be rescheduled every time after boot completes.