Add an API to provide data to `AppBottomSheet`
Most times any data required by the bottom sheet can be obtained from the screen state, but there are times in which the data comes from an item being clicked in a list.

Arguably this could also mutate the screen state, but this is a shortcut that works the same way and ends up being simpler for transient state that's only relevant while the sheet is visible.